% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DashScope.R
\name{DashScopeCellType}
\alias{DashScopeCellType}
\title{Annotates cell clusters in single-cell sequencing analysis.}
\arguments{
\item{api_key}{character. Obtained from the website https://dashscope.console.aliyun.com/apiKey.}

\item{model}{character. Check from website https://dashscope.console.aliyun.com/billing, the default value is "qwen-max-longcontext".}

\item{inputTable}{data.frame returned by FindAllMarkers.}

\item{tissueName}{The tissue origin of single-cell sequencing data, such as "mouse brain".}

\item{useGeneNumber}{numeric. The number of genes used for annotating cell types, with a default value of 10.}
}
\value{
character vector
}
\description{
This package annotates cell clusters in single-cell sequencing analysis by using ERNIEBot or DashScope models.
}
