% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERNIEBot.R
\name{ERNIEBotCellType}
\alias{ERNIEBotCellType}
\title{Annotates cell clusters in single-cell sequencing analysis.}
\arguments{
\item{account}{list. Obtained from the website https://console.bce.baidu.com/qianfan/ais/console/applicationConsole/application?_=1720524456151. client_id is the API Key, and client_secret is the Secret Key.}

\item{model}{character. Check from website https://console.bce.baidu.com/qianfan/ais/console/onlineService. the default value is "completions_pro".}

\item{inputTable}{data.frame returned by FindAllMarkers.}

\item{tissueName}{The tissue origin of single-cell sequencing data, such as "mouse brain".}

\item{useGeneNumber}{numeric. The number of genes used for annotating cell types, with a default value of 10.}
}
\value{
character vector
}
\description{
This package annotates cell clusters in single-cell sequencing analysis by using ERNIEBot or DashScope models.
}
